requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
PyMySQL==1.0.2
pycryptodome==3.12.0
python-dotenv==0.19.2
bcrypt==3.2.0
email-validator==1.1.3
Werkzeug==2.0.3